+++ xfxpal-talent/package.json
@@ -9,6 +9,7 @@
     "csvtojson": "^2.0.10",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "request": "^2.88.2",
     "semantic-ui-react": "^0.88.2",
