+++ xfxpal-talent/src/Publications.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Header } from 'semantic-ui-react'
+
+export default () => {
+    return (
+        <Header as="h1" textAlign='center' >FXPAL publishes in top scientific conferences and journals</Header>
+    )
+}
